requirements.txt
# Requirements for the Discord bot
discord.py
pytz
rich
APScheduler